robo-nexus-bot/requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
flask>=2.3.0
requests>=2.31.0
psutil>=5.9.0
psycopg2-binary>=2.9.0